updated
During my college year, I have studied in three different cities: New York, London, and Paris. And living in these cities are very similar to Drawing on the Web class in terms of dynamicity. While I was highly exposed to different cultures, I was not highly exposed to the history of the city nor the monuments. Therefore, I would like to take this chance to study the history of the city.
	The cover page of the project will be world map. The three cities will be highlighted in some way and further will have a link to a corresponding page. This page will be a simple HTML page with CSS code.
	The first page will be on Paris. This page will feature a map of Paris, with 14 top places in Paris. User will be able to see a brief history or an information on each place when they click on the image. When the image is clicked, it will create a popup window on the top of the map. The popup window will feature an icon of the place with a brief description of it. This will be done via JavaScript.
	The second page will on New York. This page will be full of HTML canvases. Each HTML canvas will have different pictures or videos of places and food that I’ve been to and I’ve had.
	The third page will be on London. I would like to incorporate WebGL to this page. When the user loads the third page, he will see the famous Big Ben. He will be also able to rotate Big Ben to explore all sides of Big Ben.